+++ plugins/content-manager/src/main/java/com/wazuh/contentmanager/settings/PluginSettings.java
@@ -38,6 +38,7 @@ public class PluginSettings {
     public static final String INTEGRATIONS_URI = PLUGINS_BASE_URI + "/integrations";
     public static final String PROMOTE_URI = PLUGINS_BASE_URI + "/promote";
     public static final String POLICY_URI = PLUGINS_BASE_URI + "/policy";
+    public static final String SPACE_URI = PLUGINS_BASE_URI + "/space";
 
     /** Settings default values */
     private static final int DEFAULT_MAX_ITEMS_PER_BULK = 999;